Using `synchronize` instead of `activate`
I noticed when run from launchd the backup takes place but Evernote fails to be quit. I see errors in the console about not being able to become the foreground application. Since becoming foreground is not needed (and indeed not wanted) I removed the call to `activate` and instead use `synchronize` to get Evernote's attention. We'll see if this allows us to quit after the backup.

I also corrected the licence.